server/requirements.txt
streamlit
streamlit_float 
openai
audio_recorder_streamlit
python-dotenv
tavily-python
crewai 
crewai[tools]
tiktoken
langchain 
langchain-openai
langchainhub
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
pvrecorder
pveagle
pvporcupine
flask_socketio
wave
flask-sqlalchemy
<<<<<<< HEAD
pvcobra
=======
flask-bcrypt
>>>>>>> f42ec63a30c79ebe8895ac58a35e7c46f3c1d46b
